warnings raus
git-svn-id: svn+ssh://iagpc151.iag.uni-stuttgart.de/usr/local/svnroot/Boltzplatz/branches/PICASSO@726 f8cfc4a6-4e3a-43c3-8588-810d72ca601e

+++ src/particles/particle_vars.f90
@@ -308,7 +308,7 @@ REAL, ALLOCATABLE                        :: vMPFOldVelo(:,:)                  ! 
 REAL, ALLOCATABLE                        :: vMPFOldBrownVelo(:,:)             ! Old brownian Velo
 REAL, ALLOCATABLE                        :: vMPFOldPos(:,:)                   ! Old Particle Pos for Polynom
 REAL, ALLOCATABLE                        :: vMPFOldMPF(:)                     ! Old Particle MPF
-REAL, ALLOCATABLE                        :: vMPFNewPosNum(:)
+INTEGER, ALLOCATABLE                     :: vMPFNewPosNum(:)
 INTEGER, ALLOCATABLE                     :: vMPF_SpecNumElem(:,:)             ! number of particles of spec (:,i) in element (j,:)
 CHARACTER(30)                            :: vMPF_velocityDistribution         ! specifying keyword for velocity distribution
 REAL, ALLOCATABLE                        :: vMPF_NewPosRefElem(:,:)          ! new positions in ref elem
